requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
aiohttp==3.9.1
tweepy==4.14.0
textblob==0.17.1
pandas==2.1.3
plotly==5.18.0
pytz==2023.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aioredis==2.0.1
backoff==2.2.1
